Bursar and Receptionist dashboard was add and some bugs fixed

+++ eportal/admin/logout.php
@@ -1,7 +1,7 @@
 <?php 
 require_once "helpers/helper.php";
 $ses_id = $_SESSION['ADMIN_TOKEN_ID'];
-if (isset($_GET['action'])) {
+if (isset($_GET['action'])  && isset($_SESSION['ADMIN_TOKEN_ID'])) {
 	if ($_GET['action'] ==="logout") {
 		if ($Admin->deleteAdminSessionToken($_SESSION['ADMIN_USERNAME'],$_SESSION['ADMIN_EMAIL'],$_SESSION['ADMIN_TOKEN'])) {
 			if ($Staff->logout($ses_id)) {
